Make current_pos a floating point number

+++ server/cdriver/cdriver.h
@@ -215,7 +215,7 @@ struct Space_History {
 struct Motor_History {
 	double last_v;		// v during last iteration, for using limit_a [m/s].
 	double target_v, target_dist;	// Internal values for moving.
-	int32_t current_pos;	// Current position of motor (in steps), and what the hardware currently thinks.
+	double current_pos;	// Current position of motor (in steps), and (cast to int) what the hardware currently thinks.
 	double endpos;
 };
 
@@ -434,7 +434,7 @@ void buffered_debug(char const *fmt, ...);
 #endif
 
 // Force cpdebug if requested, to enable only specific lines without adding all the cp things in manually.
-#define fcpdebug(s, m, fmt, ...) do { if (s == 0 && m == 0) debug("CP curfragment %d curpos %d current %f " fmt, current_fragment, spaces[s].motor[m]->settings.current_pos, spaces[s].axis[m]->settings.current, ##__VA_ARGS__); } while (0)
+#define fcpdebug(s, m, fmt, ...) do { if (s == 0 && m == 0) debug("CP curfragment %d curpos %f current %f " fmt, current_fragment, spaces[s].motor[m]->settings.current_pos, spaces[s].axis[m]->settings.current, ##__VA_ARGS__); } while (0)
 //#define cpdebug fcpdebug
 #define cpdebug(...) do {} while (0)
